longtail.h/c no longer knows compression api or compression registry
